# openapi_spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Platform God API",
    "description": "REST API for agent-driven repository governance",
    "version": "0.0.0"
  },
  "paths": {
...
    "/api/v1/runs": {
      "get": {
        "tags": [
          "Runs"
        ],
        "summary": "List Runs",
        "description": "List chain execution runs.\n\nSupports filtering by repository, chain name, and status.\n\nQuery Parameters:\n    repository_root: Filter by repository root path\n    chain_name: Filter by chain name\n    status: Filter by run status (pending, running, completed, failed, cancelled)\n    limit: Maximum results (default: 50, max: 500)\n    offset: Number of results to skip (default: 0)\n\nReturns:\n    RunListResponse with matching runs",
        "operationId": "list_runs_api_v1_runs_get",
        "parameters": [
          {
            "name": "repository_root",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by repository root",
              "title": "Repository Root"
            },
            "description": "Filter by repository root"
          },
          {
            "name": "chain_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by chain name",
              "title": "Chain Name"
            },
            "description": "Filter by chain name"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by run status",
              "title": "Status"
            },
            "description": "Filter by run status"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "description": "Maximum results",
              "default": 50,
              "title": "Limit"
            },
            "description": "Maximum results"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Results to skip",
              "default": 0,
              "title": "Offset"
            },
            "description": "Results to skip"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/recent": {
      "get": {
        "tags": [
          "Runs"
        ],
        "summary": "List Recent Runs",
        "description": "List recent chain runs across all repositories.\n\nArgs:\n    limit: Number of recent runs to return (default: 10, max: 100)\n\nReturns:\n    RunListResponse with recent runs",
        "operationId": "list_recent_runs_api_v1_runs_recent_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Number of recent runs",
              "default": 10,
              "title": "Limit"
            },
            "description": "Number of recent runs"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}": {
      "get": {
        "tags": [
          "Runs"
        ],
        "summary": "Get Run",
        "description": "Get detailed information about a specific run.\n\nArgs:\n    run_id: Unique run identifier\n\nReturns:\n    RunDetail with full run information\n\nRaises:\n    NotFoundError: If run doesn't exist",
        "operationId": "get_run_api_v1_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Runs"
        ],
        "summary": "Delete Run",
        "description": "Delete a specific run record.\n\nUseful for cleaning up old run data.\n\nArgs:\n    run_id: Unique run identifier\n\nRaises:\n    NotFoundError: If run doesn't exist",
        "operationId": "delete_run_api_v1_runs__run_id__delete",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/repository/{repository_name}": {
      "get": {
        "tags": [
          "Runs"
        ],
        "summary": "List Runs By Repository",
        "description": "List runs for a specific repository by name.\n\nArgs:\n    repository_name: Repository name (path component)\n    limit: Maximum results\n    offset: Results to skip\n\nReturns:\n    RunListResponse with matching runs\n\nRaises:\n    NotFoundError: If repository not found in any runs",
        "operationId": "list_runs_by_repository_api_v1_runs_repository__repository_name__get",
        "parameters": [
          {
            "name": "repository_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Repository Name"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/chain/{chain_name}/latest": {
      "get": {
        "tags": [
          "Runs"
        ],
        "summary": "Get Latest Run For Chain",
        "description": "Get the latest run for a specific chain and repository.\n\nArgs:\n    chain_name: Name of the chain\n    repository_root: Repository root path\n\nReturns:\n    RunDetail with latest run information\n\nRaises:\n    NotFoundError: If no run found",
        "operationId": "get_latest_run_for_chain_api_v1_runs_chain__chain_name__latest_get",
        "parameters": [
          {
            "name": "chain_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Chain Name"
            }
          },
          {
            "name": "repository_root",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Repository root path",
              "title": "Repository Root"
            },
            "description": "Repository root path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RunDetail": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id",
            "description": "Unique run identifier"
          },
          "chain_name": {
            "type": "string",
            "title": "Chain Name",
            "description": "Chain name"
          },
          "repository_root": {
            "type": "string",
            "title": "Repository Root",
            "description": "Repository root path"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus",
            "description": "Run status"
          },
          "started_at": {
            "type": "string",
            "title": "Started At",
            "description": "Start timestamp"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At",
            "description": "Completion timestamp"
          },
          "execution_time_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Execution Time Ms",
            "description": "Execution time"
          },
          "agent_results": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Agent Results",
            "description": "Agent results"
          },
          "final_state": {
            "additionalProperties": true,
            "type": "object",
            "title": "Final State",
            "description": "Final state"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error",
            "description": "Error message if failed"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run_id",
          "chain_name",
          "repository_root",
          "status",
          "started_at"
        ],
        "title": "RunDetail",
        "description": "Detailed information for a chain run."
      },
      "RunListResponse": {
        "properties": {
          "runs": {
            "items": {
              "$ref": "#/components/schemas/RunSummary"
            },
            "type": "array",
            "title": "Runs",
            "description": "List of run summaries"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Total count matching filters"
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "description": "Response limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset",
            "description": "Response offset"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "runs",
          "total",
          "limit",
          "offset"
        ],
        "title": "RunListResponse",
        "description": "Response model for listing runs."
      },
      "RunStatus": {
        "type": "string",
        "enum": [
          "pending",
          "running",
          "completed",
          "failed",
          "cancelled"
        ],
        "title": "RunStatus",
        "description": "Status of a chain run."
      },
      "RunSummary": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id",
            "description": "Unique run identifier"
          },
          "chain_name": {
            "type": "string",
            "title": "Chain Name",
            "description": "Chain name"
          },
          "repository": {
            "type": "string",
            "title": "Repository",
            "description": "Repository name"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus",
            "description": "Run status"
          },
          "started_at": {
            "type": "string",
            "title": "Started At",
            "description": "Start timestamp"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At",
            "description": "Completion timestamp"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms",
            "description": "Duration in milliseconds"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run_id",
          "chain_name",
          "repository",
          "status",
          "started_at"
        ],
        "title": "RunSummary",
        "description": "Summary information for a chain run."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}